requirement.txt
cycler==0.10.0
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.17.2
opencv-python==4.1.1.26
pyparsing==2.4.2
python-dateutil==2.8.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
wincertstore==0.2
elasticsearch
six
line-bot-sdk
werkzeug>=0.15.4